+++ src/ngrest/base/actions/UpdateAction.php
@@ -4,6 +4,8 @@ namespace luya\admin\ngrest\base\actions;
 
 use Yii;
 use luya\admin\models\UserOnline;
+use yii\web\ServerErrorHttpException;
+use yii\web\NotFoundHttpException;
 
 /**
  * UpdateAction for REST implementation.
@@ -15,10 +17,24 @@ class UpdateAction extends \yii\rest\UpdateAction
 {
     public function run($id)
     {
-        $response = parent::run($id);
+        /* @var $model ActiveRecord */
+        $model = $this->controller->findModelClassObject($this->modelClass, $id, 'update');
+
+        if (!$model) {
+            throw new NotFoundHttpException("Object not found: $id");
+        }
         
+        if ($this->checkAccess) {
+            call_user_func($this->checkAccess, $this->id, $model);
+        }
+        $model->scenario = $this->scenario;
+        $model->load(Yii::$app->getRequest()->getBodyParams(), '');
+        if ($model->save() === false && !$model->hasErrors()) {
+            throw new ServerErrorHttpException('Failed to update the object for unknown reason.');
+        }
+
         UserOnline::unlock(Yii::$app->adminuser->id);
-        
-        return $response;
+
+        return $model;
     }
 }