+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This release contains new migrations and requires to run the `migrate` command a
 
 ## Changed
 
++ [#261](https://github.com/luyadev/luya-module-admin/issues/261) Add ngRestFind() for none $is_api_user Users.
 + [#210](https://github.com/luyadev/luya-module-admin/issues/210) New tag translation option.
 + [#140](https://github.com/luyadev/luya-module-admin/issues/140) Generic Scheduler with Yii Queue integration.